HUB-737: Change to use Idps available for registration
This step is to decouple from the weirdness of variant_partial_controller
and the use of special-cases folder
