+++ app/controllers/failed_registration_loa2_controller.rb
@@ -7,7 +7,7 @@ class FailedRegistrationLoa2Controller < ApplicationController
     @idp = IDENTITY_PROVIDER_DISPLAY_DECORATOR.decorate(selected_identity_provider)
     @transaction = current_transaction
     @custom_fail = current_transaction.custom_fail_heading.present?
-    @idp_recommendation_engine = IDP_RECOMMENDATION_ENGINE_variant_c
+    @idp_recommendation_engine = IDP_RECOMMENDATION_ENGINE
     render view_for_loa2
   end
