+++ app/controllers/partials/variant_partial_controller.rb
@@ -1,6 +1,6 @@
 module VariantPartialController
-  def current_identity_providers_for_loa_by_variant(variant)
-    current_available_identity_providers_for_registration.select { |idp| ABC_VARIANTS_CONFIG["variant_#{variant}_idp_set"].include?(idp.simple_id) }
+  def current_identity_providers_for_loa_by_variant(_variant)
+    current_available_identity_providers_for_registration
   end
 
   def segment_advice(segments)
