+++ app/views/choose_a_certified_company/choose_a_certified_company_LOA2.html.erb
@@ -1,49 +1,45 @@
-<%= page_title 'hub.choose_a_certified_company.title' %>
+<%= page_title 'hub_variant_c.choose_a_certified_company.title' %>
 <% content_for :feedback_source, 'CHOOSE_A_CERTIFIED_COMPANY_PAGE' %>
 <% content_for :additional_javascript do %>
     <%= javascript_include_tag 'piwik_idp_picker_tracking' %>
 <% end %>
+
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l"><%= t 'hub.choose_a_certified_company.heading' %></h1>
-    <p><%= t 'hub.choose_a_certified_company.personal_information' %></p>
-    <ul class="govuk-list govuk-list--bullet">
-      <li><%= link_to(t('hub.choose_a_certified_company.why_companies'), why_companies_path) %></li>
-    </ul>
-    <p class="govuk-body">
-      <%= t 'hub.choose_a_certified_company.idp_count_html',
-            company_count: t('hub.choose_a_certified_company.company_html', count: @recommended_idps.length) %>
-    </p>
-
-
-    <div id="matching-idps">
-      <% @recommended_idps.each.with_index do |identity_provider, index| %>
-        <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: true, identity_provider: identity_provider, order: index + 1} %>
-      <% end %>
+      <h1 class="govuk-heading-l"><%= t 'hub_variant_c.choose_a_certified_company.heading' %></h1>
+      <p><%= t 'hub_variant_c.choose_a_certified_company.intro' %></p>
+      <p>
+        <%= t 'hub_variant_c.choose_a_certified_company.idp_count' %>
+        <%= t 'hub_variant_c.choose_a_certified_company.idp_explanation' %>
+      </p>
+      <%= t 'hub_variant_c.choose_a_certified_company.content_html' %>
+      <div id="matching-idps">
+        <% @recommended_idps.each.with_index do |identity_provider, index| %>
+          <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: true, identity_provider: identity_provider, order: index + 1} %>
+        <% end %>
+      </div>
     </div>
 
-    <% if @non_recommended_idps.any? %>
-    
-      <div id="non-matching-idps">
-        <div class="grid-row">
-          <div class="column-two-thirds govuk-!-padding-top-3">
-            <p class="govuk-body">
-              <%= t 'hub.choose_a_certified_company.filtered_idps_message_html',
-                    company_count: t('hub.choose_a_certified_company.company_html', count: @non_recommended_idps.length) %>
-            </p>
-          </div>
+  <% if @non_recommended_idps.any? %>
+    <div id="non-matching-idps">
+      <div class="grid-row">
+        <div class="column-two-thirds govuk-!-padding-top-3">
+          <p class="govuk-body">
+            <%= t 'hub.choose_a_certified_company.filtered_idps_message_html',
+                  company_count: t('hub.choose_a_certified_company.company_html', count: @non_recommended_idps.length) %>
+          </p>
         </div>
-        <details class="govuk-details" data-module="govuk-details">
-          <summary id="non-matching-idps-trigger" class="govuk-details__summary">
-            <span class="govuk-details__summary-text"><%= t 'hub.choose_a_certified_company.show_all_companies' %></span>
-          </summary>
-          <div class="govuk-details__text">
-            <% @non_recommended_idps.each.with_index do |identity_provider, index| %>
-              <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: false, identity_provider: identity_provider, order: index + 1} %>
-            <% end %>
-          </div>
-        </details>
       </div>
-    <% end %>
-  </div>
+      <details class="govuk-details" data-module="govuk-details">
+        <summary id="non-matching-idps-trigger" class="govuk-details__summary">
+          <span class="govuk-details__summary-text"><%= t 'hub.choose_a_certified_company.show_all_companies' %></span>
+        </summary>
+        <div class="govuk-details__text">
+          <% @non_recommended_idps.each.with_index do |identity_provider, index| %>
+            <%= render partial: 'choose_a_certified_company/idp_option', locals: {recommended: false, identity_provider: identity_provider, order: index + 1} %>
+          <% end %>
+        </div>
+      </details>
+    </div>
+  <% end %>
 </div>
