+++ spec/controllers/about_loa2_controller_spec.rb
@@ -8,9 +8,7 @@ describe AboutLoa2Controller do
 
   before(:each) do
     stub_request(:get, CONFIG.config_api_host + "/config/transactions/enabled")
-    experiment = "short_hub_2019_q3"
-    variant = "variant_c_2_idp_short_hub"
-    set_session_and_cookies_with_loa_and_variant("LEVEL_2", experiment, variant)
+    set_session_and_cookies_with_loa("LEVEL_2")
     stub_api_idp_list_for_registration
   end
 
@@ -22,7 +20,7 @@ describe AboutLoa2Controller do
     end
 
     it "renders the LOA2 certified companies for variant C on the combined about view" do
-      expect(identity_provider_display_decorator).to receive(:decorate_collection).with(a_list_of_size(2)).and_return([])
+      expect(identity_provider_display_decorator).to receive(:decorate_collection).with(a_list_of_size(7)).and_return([])
       expect(subject).to render_template(:about_combined_LOA2)
     end
   end
