--- stub/idp-rules-variant-c/stub-idp-one.yml
@@ -1,14 +1,0 @@
-simpleIds: [ "stub-idp-one" ]
-send_hints: true
-send_language_hint: true
-
-segments:
-  protected:
-    recommended: ["pp_app", "dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
-    unlikely:    []
-  non_protected:
-    recommended: ["pp_app", "dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
-    unlikely:    []
-
-capabilities:
-  - []
