--- stub/idp-rules-variant-c/stub-idp-three.yml
@@ -1,12 +1,0 @@
-simpleIds: [ "stub-idp-three" ]
-
-segments:
-  protected:
-    recommended: ["dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
-    unlikely:    ["pp_app"]
-  non_protected:
-    recommended: ["dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
-    unlikely:    ["pp_app"]
-
-capabilities:
-  - []
