+++ stub/idp-rules/stub-idp-one.yml
@@ -4,11 +4,11 @@ send_language_hint: true
 
 segments:
   protected:
-    recommended: ["other"]
-    unlikely: []
+    recommended: ["pp_app", "dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
+    unlikely:    []
   non_protected:
-    recommended: ["protected-segment-1", "non-protected-segment-3"]
-    unlikely: ["non-protected-segment-2"]
+    recommended: ["pp_app", "dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
+    unlikely:    []
 
 capabilities:
   - []
