+++ stub/idp-rules/stub-idp-three.yml
@@ -2,11 +2,11 @@ simpleIds: [ "stub-idp-three" ]
 
 segments:
   protected:
-    recommended: []
-    unlikely: []
+    recommended: ["dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
+    unlikely:    ["pp_app"]
   non_protected:
-    recommended: ["non-protected-segment-4"]
-    unlikely: []
+    recommended: ["dl_app_cc", "pp_app_cc", "pp_dl_cc", "pp_dl_app", "pp_dl_app_cc"]
+    unlikely:    ["pp_app"]
 
 capabilities:
   - []
