--- stub/segment_definitions_variant_c.yml
@@ -1,67 +1,0 @@
-segments:
-- "0doc"
-- "cc"
-- "app"
-- "app_cc"
-- "dl"
-- "dl_cc"
-- "dl_app"
-- "dl_app_cc"
-- "pp"
-- "pp_cc"
-- "pp_app"
-- "pp_app_cc"
-- "pp_dl"
-- "pp_dl_cc"
-- "pp_dl_app"
-- "pp_dl_app_cc"
-
-segment_definitions:
-  0doc:
-    - ["device_type_other"]
-    - ["device_type_mobile"]
-  cc:
-    - ["device_type_other", "has_credit_card"]
-    - ["device_type_mobile", "has_credit_card"]
-  app:
-    - ["device_type_other", "has_phone_can_app"]
-    - ["device_type_mobile", "has_phone_can_app"]
-  app_cc:
-    - ["device_type_other", "has_phone_can_app", "has_credit_card"]
-    - ["device_type_mobile", "has_phone_can_app", "has_credit_card"]
-  dl:
-    - ["device_type_other", "has_driving_license"]
-    - ["device_type_mobile", "has_driving_license"]
-  dl_cc:
-    - ["device_type_other", "has_driving_license", "has_credit_card"]
-    - ["device_type_mobile", "has_driving_license", "has_credit_card"]
-  dl_app:
-    - ["device_type_other", "has_driving_license", "has_phone_can_app"]
-    - ["device_type_mobile", "has_driving_license", "has_phone_can_app"]
-  dl_app_cc:
-    - ["device_type_other", "has_driving_license", "has_phone_can_app", "has_credit_card"]
-    - ["device_type_mobile", "has_driving_license", "has_phone_can_app", "has_credit_card"]
-  pp:
-    - ["device_type_other", "has_valid_passport"]
-    - ["device_type_mobile", "has_valid_passport"]
-  pp_cc:
-    - ["device_type_other", "has_valid_passport", "has_credit_card"]
-    - ["device_type_mobile", "has_valid_passport", "has_credit_card"]
-  pp_app:
-    - ["device_type_other", "has_valid_passport", "has_phone_can_app"]
-    - ["device_type_mobile", "has_valid_passport", "has_phone_can_app"]
-  pp_app_cc:
-    - ["device_type_other", "has_valid_passport", "has_phone_can_app", "has_credit_card"]
-    - ["device_type_mobile", "has_valid_passport", "has_phone_can_app", "has_credit_card"]
-  pp_dl:
-    - ["device_type_other", "has_valid_passport", "has_driving_license"]
-    - ["device_type_mobile", "has_valid_passport", "has_driving_license"]
-  pp_dl_cc:
-    - ["device_type_other", "has_valid_passport", "has_driving_license", "has_credit_card"]
-    - ["device_type_mobile", "has_valid_passport", "has_driving_license", "has_credit_card"]
-  pp_dl_app:
-    - ["device_type_other", "has_valid_passport", "has_driving_license", "has_phone_can_app"]
-    - ["device_type_mobile", "has_valid_passport", "has_driving_license", "has_phone_can_app"]
-  pp_dl_app_cc:
-    - ["device_type_other", "has_valid_passport", "has_driving_license", "has_phone_can_app", "has_credit_card"]
-    - ["device_type_mobile", "has_valid_passport", "has_driving_license", "has_phone_can_app", "has_credit_card"]
